Unify behaviour regarding virtual remotes and add query parameter (#11)
Previously devicelist.cgi and statelist.cgi handled virtual remotes differently. devicelist.cgi ommited these channels based on the type of the device, statelist.cgi used the name of the device to achieve the same, which yielded inconsitent behaviour. Now the device type is used in both scripts. This breaks the possiblity to receive the workaround to receive the virtual remote in statelist.cgi by renaming the device.

Therefore this also adds a query parameter show_remote to both scripts to reliably receive the virtual remote channels.

Co-authored-by: Jens Maus <mail@jens-maus.de>